feat(oop): drastically change type system
It's now only type ids, and objects store their kinds. This change also reduced object header size from 24 to 16 bytes.

+++ framework/include/BibbleVM/core/oop/class.h
@@ -54,7 +54,7 @@ namespace bibblevm::oop {
         uint16_t mObjectFieldCount;
 
         void sortFields();
-        void orderFieldBuckets(std::array<FieldBucket, Type::Count>& buckets);
+        void orderFieldBuckets(std::array<FieldBucket, static_cast<size_t>(Type::Count)>& buckets);
         static void orderFieldBucket(FieldBucket& bucket, uint32_t size, uint64_t& offset);
 
         void createVtable();
